+++ libft/src/ft_strrchr.c
@@ -16,7 +16,7 @@ char	*ft_strrchr(const char *s, int c)
 {
 	int		i;
 
-	i = ft_strlen(s);
+	i = (int)ft_strlen(s);
 	while (i >= 0 && s[i] != (char)c)
 		i--;
 	return (i < 0 ? (char *)NULL : ((char *)&s[i]));
