+++ libft/src/ft_strtrim.c
@@ -20,13 +20,13 @@ char	*ft_strtrim(char const *s)
 	if (s)
 	{
 		i = 0;
-		j = ft_strlen(s) - 1;
+		j = (int)ft_strlen(s) - 1;
 		while (s[i] == ' ' || s[i] == '\n' || s[i] == '\t')
 			i++;
 		while (s[j] == ' ' || s[j] == '\n' || s[j] == '\t')
 			j--;
 		if (j >= i)
-			return (ft_strsub(s, i, (j - i + 1)));
+			return (ft_strsub(s, (unsigned)i, (size_t)j - i + 1));
 		else
 			return (ft_strnew(0));
 	}
